Multiple changes.
 * Render first code in proper code block.
 * Properly mark hunts by system users (incl. GRRCron) as robot hunts.
 * Use the `URL` interface for body export URL rather than a raw string.
 * Add support for body export options in the new timeline flow UI.
 * Add support for progress updates in the timeline flow.
 * Extend the timeline action with reports about the number of scanned entries.
 * Create an iterator wrapper utility that counts the number of returned entries.
 * Adjust the way the timeline action reports results.
 * Display basic results for ListProcesses.
 * Not outputting logs during tests.
 * Fix loading of libc on macOS Big Sur.
 * Write an end-to-end test for body export options.
 * Autocomplete artifacts based on description and sources (e.g. paths).
 * Show artifacts unavailable on the current OS as gray.
 * Create an API integration test for timeline methods.
 * Move timeline creation test utility to a separate module.
 * Add support for body export options in the gzc2body tool.
 * Refactor the body export utility to use a dedicated class for options.
 * Add support for body export options in the API client.
 * Add support for nested messages in GET requests.
 * Show one-liner preview for all artifact source types.
 * Introduced Hunt.default_client_rate config parameter (set to 200 for prod/staging).
 * Show artifact description and source in autocomplete.
 * Disable dynamic type lookup for hunt arguments.
 * Create an API value renderer for Protcol Buffers `Any` messages.
 * Create an auxiliary method for unpacking from RDF wrapper of `Any`.
 * Require flow cron jobs to always specify flow name.
 * Add a convenience constructor for packing RDF values into `Any`.
 * Add support for disabling dynamic type lookup in `Any` fields.
 * Remove an unused assembly file.
 * Removing "unclosed file" warnings from Python tests.
 * Show approver suggestions for the empty input field.
 * Suggest previously requested approvers initially.
 * Sandboxing: integrate Windows process management into communication module.
 * Remove a no longer used constant.
 * Fix distinctUntilChanged() with Arrays and Objects.
 * Autocomplete artifacts in artifact flow form.

+++ grr/client/grr_response_client/unprivileged/communication.py
@@ -2,10 +2,12 @@
 """Communication protocol with unprivileged servers."""
 
 import abc
-import socket
+import contextlib
+import os
+import platform
 import struct
 import subprocess
-from typing import NamedTuple, Callable, Optional, List
+from typing import NamedTuple, Callable, Optional, List, BinaryIO
 
 
 class Transport(abc.ABC):
@@ -32,7 +34,7 @@ class Message(NamedTuple):
   """Additional data contained in a message, usually just raw bytes."""
 
 
-_HEADER_STRUCT = struct.Struct('<LL')
+_HEADER_STRUCT = struct.Struct("<LL")
 
 
 class Connection:
@@ -64,11 +66,11 @@ class Connection:
     if data_len > 0:
       data = self._transport.RecvBytes(data_len)
     else:
-      data = b''
+      data = b""
     if attachment_len > 0:
       attachment = self._transport.RecvBytes(attachment_len)
     else:
-      attachment = b''
+      attachment = b""
     return Message(data=data, attachment=attachment)
 
 
@@ -91,28 +93,46 @@ class Server(abc.ABC):
     pass
 
 
-class SocketTransport(Transport):
-  """A low-level transport protocol based on a socket."""
-
-  def __init__(self, sock: socket.socket):
-    self._socket = sock
-
-  def SendBytes(self, data: bytes) -> None:
-    self._socket.sendall(data)
-
-  def RecvBytes(self, size: int) -> bytes:
-    return self._socket.recv(size, socket.MSG_WAITALL)
-
-
 ConnectionHandler = Callable[[Connection], None]
 
-ArgsFactory = Callable[[int], List[str]]
+
+class Channel(NamedTuple):
+  """File descriptors / handle sused for communication."""
+
+  pipe_input: Optional[int] = None
+  """Pipe used for input (client to server).
+
+  This is a file descriptor on UNIX, a handle on Windows.
+  """
+
+  pipe_output: Optional[int] = None
+  """Pipe used for output (server to client).
+
+  This is a file descriptor on UNIX, a handle on Windows.
+  """
+
+
+ArgsFactory = Callable[[Channel], List[str]]
+
+
+class PipeTransport(Transport):
+  """A low-level transport protocol based on a pair of pipes."""
+
+  def __init__(self, read_pipe: BinaryIO, write_pipe: BinaryIO):
+    self._read_pipe = read_pipe
+    self._write_pipe = write_pipe
+
+  def SendBytes(self, data: bytes) -> None:
+    self._write_pipe.write(data)
+
+  def RecvBytes(self, size: int) -> bytes:
+    return self._read_pipe.read(size)
 
 
 class SubprocessServer(Server):
   """A server running as a subprocess.
 
-  A socketpair is created and shared with the subprocess as communication
+  A pair of pipes is created and shared with the subprocess as communication
   channel.
   """
 
@@ -120,42 +140,81 @@ class SubprocessServer(Server):
     """Constructor.
 
     Args:
-      args_factory: Function which takes a socket file descriptor and returns
+      args_factory: Function which takes a channel and returns
         the args to run the server subprocess (as required by subprocess.Popen).
     """
     self._args_factory = args_factory
     self._process = None  # type: Optional[subprocess.Popen]
-    self._socket = None  # type: Optional[socket.socket]
-    self._remote_socket = None  # type: Optional[socket.socket]
+    # Omiting type, since the type is conditionally imported on Windows.
+    self._process_win = None
+    self._output_r = None  # type: Optional[BinaryIO]
+    self._input_w = None  # type: Optional[BinaryIO]
 
   def Start(self) -> None:
-    self._socket, self._remote_socket = socket.socketpair()
-    args = self._args_factory(self._remote_socket.fileno())
-    self._process = subprocess.Popen(
-        args,
-        close_fds=True,
-        pass_fds=[self._remote_socket.fileno()],
-    )
-    self._remote_socket.close()
+    with contextlib.ExitStack() as stack:
+      input_r_fd, input_w_fd = os.pipe()
+      stack.callback(os.close, input_r_fd)
+      self._input_w = os.fdopen(input_w_fd, "wb", buffering=0)
+
+      output_r_fd, output_w_fd = os.pipe()
+      stack.callback(os.close, output_w_fd)
+      self._output_r = os.fdopen(output_r_fd, "rb", buffering=0)
+
+      if platform.system() == "Windows":
+        # pylint: disable=g-import-not-at-top
+        import msvcrt
+        from grr_response_client.unprivileged.windows import process  # pytype: disable=import-error
+        # pylint: enable=g-import-not-at-top
+        # pytype doesn't see the functions in msvcrt
+        # pytype: disable=module-attr
+        input_r_handle = msvcrt.get_osfhandle(input_r_fd)
+        output_w_handle = msvcrt.get_osfhandle(output_w_fd)
+        # pytype: enable=module-attr
+        args = self._args_factory(
+            Channel(pipe_input=input_r_handle, pipe_output=output_w_handle))
+        self._process_win = process.Process(args, [input_r_fd, output_w_fd])
+      else:
+        args = self._args_factory(
+            Channel(pipe_input=input_r_fd, pipe_output=output_w_fd))
+        self._process = subprocess.Popen(
+            args,
+            close_fds=True,
+            pass_fds=[input_r_fd, output_w_fd],
+        )
 
   def Stop(self) -> None:
-    self._process.kill()
-    self._process.wait()
-    self._socket.close()
+    if self._process is not None:
+      self._process.kill()
+      self._process.wait()
+    if self._process_win is not None:
+      self._process_win.Stop()
+    if self._input_w is not None:
+      self._input_w.close()
+    if self._output_r is not None:
+      self._output_r.close()
 
   def Connect(self) -> Connection:
-    transport = SocketTransport(self._socket)
+    transport = PipeTransport(self._output_r, self._input_w)
     return Connection(transport)
 
 
-def Main(socket_fd: int, connection_handler: ConnectionHandler) -> None:
+def Main(channel: Channel, connection_handler: ConnectionHandler) -> None:
   """The entry point of the server process.
 
   Args:
-    socket_fd: Socket file descriptor connected to the client.
+    channel: Channel connected to the client.
     connection_handler: Connection handler for processing the connection.
   """
-  sock = socket.socket(fileno=socket_fd)
-  transport = SocketTransport(sock)
-  connection = Connection(transport)
-  connection_handler(connection)
+  if platform.system() == "Windows":
+    import msvcrt  # pylint: disable=g-import-not-at-top
+    # pytype doesn't see the functions in mscvcrt
+    # pytype: disable=module-attr
+    channel = Channel(
+        pipe_input=msvcrt.open_osfhandle(channel.pipe_input, os.O_RDONLY),
+        pipe_output=msvcrt.open_osfhandle(channel.pipe_output, os.O_APPEND))
+    # pytype: enable=module-attr
+  with os.fdopen(channel.pipe_input, "rb", buffering=False) as pipe_input:
+    with os.fdopen(channel.pipe_output, "wb", buffering=False) as pipe_output:
+      transport = PipeTransport(pipe_input, pipe_output)
+      connection = Connection(transport)
+      connection_handler(connection)
